chore: move `cache` mod to the root and rename it to `storage`
