+++ src/async_fuse/memfs/mod.rs
@@ -1,5 +1,4 @@
 //! The implementation of user space file system
-pub mod cache;
 mod fs_util;
 pub mod id_alloc;
 mod id_alloc_used;
@@ -33,7 +32,6 @@ pub use s3_metadata::S3MetaData;
 use serde::{Deserialize, Serialize};
 use tracing::{debug, error, info, warn};
 
-use self::cache::StorageManager;
 use self::kv_engine::KVEngineType;
 use crate::async_fuse::fuse::file_system::FileSystem;
 use crate::async_fuse::fuse::fuse_reply::{
@@ -45,6 +43,7 @@ use crate::async_fuse::fuse::protocol::{INum, FUSE_ROOT_ID};
 use crate::async_fuse::memfs::metadata::ReqContext;
 use crate::async_fuse::util::build_error_result_from_errno;
 use crate::common::error::{Context, DatenLordResult};
+use crate::storage::StorageManager;
 
 /// In-memory file system
 #[derive(Debug)]
