+++ src/storage/memory_cache/cache.rs
@@ -11,8 +11,8 @@ use tracing::warn;
 
 use super::write_back_task::Command;
 use crate::async_fuse::fuse::protocol::INum;
-use crate::async_fuse::memfs::cache::policy::EvictPolicy;
-use crate::async_fuse::memfs::cache::{Block, BlockCoordinate, BlockId, Storage};
+use crate::storage::policy::EvictPolicy;
+use crate::storage::{Block, BlockCoordinate, BlockId, Storage};
 
 /// Merge the content from `src` to `dst`. This will set `dst` to be dirty.
 fn merge_two_blocks(src: &Block, dst: &mut Block) {
