+++ src/storage/memory_cache/tests.rs
@@ -7,9 +7,9 @@ use std::time::Duration;
 use datenlord::config::SoftLimit;
 
 use super::{MemoryCache, MemoryCacheBuilder};
-use crate::async_fuse::memfs::cache::mock::MemoryStorage;
-use crate::async_fuse::memfs::cache::policy::LruPolicy;
-use crate::async_fuse::memfs::cache::{Block, BlockCoordinate, Storage};
+use crate::storage::mock::MemoryStorage;
+use crate::storage::policy::LruPolicy;
+use crate::storage::{Block, BlockCoordinate, Storage};
 
 const BLOCK_SIZE_IN_BYTES: usize = 8;
 const BLOCK_CONTENT: &[u8; BLOCK_SIZE_IN_BYTES] = b"foo bar ";
